pyproject.toml
[tool.poetry]
name = "session-scraper"
version = "0.1.0"
description = ""
authors = ["Mathijs de Bruin <mathijs@mathijsfietst.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pydantic = "^2.7.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ipython = "^8.25.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
